--- src/components/Sign/SignUpForm.tsx
@@ -1,3 +1,0 @@
-import { useForm } from 'react-hook-form';
-
-export default function SignUpForm() {}
