+++ src/styles/signForm.module.css
@@ -1,27 +1,3 @@
-.signBackground {
-  width: 100vw;
-  height: 100vh;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background: var(--Linkbrary-bg, #f0f6ff);
-}
-
-.formContainer {
-  display: flex;
-  width: 25rem;
-  padding: 14.875rem 0 15.75rem;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 32px;
-  @media screen and (max-width: 767px) {
-    width: 100%;
-    padding: 7.5rem 2rem 14.5625rem 2rem;
-    gap: 1.875rem;
-  }
-}
-
 .signForm {
   width: 100%;
   display: flex;
@@ -31,7 +7,8 @@
 }
 
 .email,
-.password {
+.password,
+.passwordConfirm {
   width: 100%;
   position: relative;
   max-width: 25rem;
@@ -39,6 +16,7 @@
   flex-direction: column;
   gap: 12px;
 }
+
 .errorMessage {
   color: var(--Linkbrary-red, #ff5b56);
   font-size: 14px;
